Reorganize fixutres to move potentially reusable ones to package level testing conftest.py.  Also modify type checking imports to avoid circular imports
